ml-api/requirements.txt
flask
joblib
scikit-learn
numpy
tensorflow
pillow